Changing the create test destination to tests.py
